requirements.txt
# ============================================
# FHP Detection System - Dependencies
# ============================================

# Core ML
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
numpy>=1.24.0
scipy>=1.10.0

# Pose Estimation
mediapipe>=0.10.0
opencv-python>=4.8.0
onnxruntime>=1.15.0

# Data Science
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Training Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
optuna>=3.2.0
pyyaml>=6.0

# Data Collection
yt-dlp>=2023.7.6
Pillow>=10.0.0
requests>=2.31.0

# Labeling
gradio>=3.40.0

# Real-time App
PyQt5>=5.15.0
playsound>=1.3.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Export
onnx>=1.14.0
